resnet-50-cifar100-custom/requirements.txt
transformers
pillow
datasets
evaluate
torch
torchvision
scikit-learn
evaluate
accelerate
tensorboardX
pandas
matplotlib